[Refactor] Refactor Update cluster
